pyproject.toml
[project]
name = "dataform-pre-commit-hooks"
version = "0.0.5"
description = "pre-commit hooks for Dataform"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "google-cloud-bigquery>=3.40.1",
    "sqlfluff",
]

[project.scripts]
dataform-sqlfluff-lint = "dataform_sqlfluff_lint:main"
dataform-syntax-check = "dataform_bigquery_emulator_syntax_check:main"

[tool.setuptools]
py-modules = ["dataform_sqlfluff_lint", "dataform_bigquery_emulator_syntax_check"]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]
